Consolidate Zcash network selection
