+++ demo-app/src/main/java/cash/z/ecc/android/sdk/demoapp/WalletCoordinator.kt
@@ -3,7 +3,7 @@ package cash.z.ecc.android.sdk.demoapp
 import android.content.Context
 import cash.z.ecc.android.sdk.Synchronizer
 import cash.z.ecc.android.sdk.demoapp.model.PersistableWallet
-import cash.z.ecc.android.sdk.demoapp.type.fromResources
+import cash.z.ecc.android.sdk.demoapp.util.fromResources
 import cash.z.ecc.android.sdk.demoapp.util.Twig
 import cash.z.ecc.android.sdk.ext.onFirst
 import cash.z.ecc.android.sdk.model.LightWalletEndpoint
