--- demo-app/src/main/res/values/bools.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- Resource that can be overlaid downstream with build variants to support
-     testnet builds. -->
-    <bool name="zcash_is_testnet">false</bool>
-</resources>
